refactor(core): split vm types and tests
